Moved a large number of legacy files under abyss/legacy directory to avoid confusion
