+++ abyss/src/abyss/legacy/cleanup_template.py
