+++ abyss/src/abyss/legacy/holes_3d_plot.py
