+++ abyss/src/abyss/legacy/parse_fastener.py
