+++ abyss/src/abyss/legacy/rolling_gradient.py
