+++ abyss/src/abyss/legacy/signal_template.py
